Move down getPageList to place of usage
